testing of PI loop
working ok needs rpm2pwr to be more accurate

+++ src/Robot/MotorControl.h
@@ -82,7 +82,7 @@ public:
   int Percent2RPM(float pct);
   float RPM2Percent(int rpm);
 
-  float ramp(float requestedPower, float accelRate);
+  int ramp(int requestedPower, float accelRate);
 
   // Encoder Related Functions
   void readEncoder();
